refactor(metadata): decompose acl_windows.rs into submodules (SPL-15 #2437)
