+++ crates/metadata/src/acl_windows/tests/mod.rs
@@ -1,0 +1,9 @@
+//! Test modules for the Windows ACL submodules, grouped by concern so
+//! each file stays focused on a single area of behaviour.
+
+mod common;
+mod dacl;
+mod posix_map;
+mod sddl;
+mod sync;
+mod xattr;
